+++ packages/brick_offline_first/lib/mixins.dart
@@ -1,2 +1,3 @@
 export 'package:brick_offline_first/src/mixins/delete_all_mixin.dart';
 export 'package:brick_offline_first/src/mixins/destructive_local_sync_from_remote_mixin.dart';
+export 'package:brick_offline_first/src/mixins/get_first_mixin.dart';