+++ packages/brick_offline_first_with_rest/lib/src/offline_first_with_rest_repository.dart
@@ -14,13 +14,8 @@ import 'package:meta/meta.dart';
 /// is unable to make contact with the [remoteProvider], the queue automatically retries in
 /// sequence until it receives a response. Please note that a response may still be an error
 /// code such as `404` or `500`. The queue is **only** concerned with connectivity.
-///
-/// OfflineFirstWithRestRepository should accept a type argument such as
-/// <_RepositoryModel extends OfflineFirstWithRestModel>, however, this causes a type bound
-/// error on runtime. The argument should be reintroduced with a future version of the
-/// compiler/analyzer.
-abstract class OfflineFirstWithRestRepository
-    extends OfflineFirstRepository<OfflineFirstWithRestModel> {
+abstract class OfflineFirstWithRestRepository<TRepositoryModel extends OfflineFirstWithRestModel>
+    extends OfflineFirstRepository<TRepositoryModel> {
   /// The type declaration is important here for the rare circumstances that
   /// require interfacting with [RestProvider]'s client directly.
   @override
@@ -100,7 +95,7 @@ abstract class OfflineFirstWithRestRepository
   }
 
   @override
-  Future<bool> delete<TModel extends OfflineFirstWithRestModel>(
+  Future<bool> delete<TModel extends TRepositoryModel>(
     TModel instance, {
     OfflineFirstDeletePolicy policy = OfflineFirstDeletePolicy.optimisticLocal,
     Query? query,
@@ -120,9 +115,9 @@ abstract class OfflineFirstWithRestRepository
   }
 
   @override
-  Future<List<TModel>> get<TModel extends OfflineFirstWithRestModel>({
+  Future<List<TModel>> get<TModel extends TRepositoryModel>({
     OfflineFirstGetPolicy policy = OfflineFirstGetPolicy.awaitRemoteWhenNoneExist,
-    query,
+    Query? query,
     bool seedOnly = false,
   }) async {
     try {
@@ -165,7 +160,7 @@ abstract class OfflineFirstWithRestRepository
   /// [OfflineFirstException] for responses that include a code within `reattemptForStatusCodes`.
   /// Defaults `false`.
   @override
-  Future<TModel> upsert<TModel extends OfflineFirstWithRestModel>(
+  Future<TModel> upsert<TModel extends TRepositoryModel>(
     TModel instance, {
     OfflineFirstUpsertPolicy policy = OfflineFirstUpsertPolicy.optimisticLocal,
     Query? query,
@@ -186,7 +181,7 @@ abstract class OfflineFirstWithRestRepository
 
   @protected
   @override
-  Future<List<TModel>> hydrate<TModel extends OfflineFirstWithRestModel>({
+  Future<List<TModel>> hydrate<TModel extends TRepositoryModel>({
     bool deserializeSqlite = true,
     Query? query,
   }) async {